Aggiunta classe customer e funzioni CRUD

+++ Cloud-Dottor/ITS.Vendrame.IITS.Northwind/ITS.Vendrame.IITS.Northwind.WebApi/Controllers/CategoriesController.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using ITS.Vendrame.IITS.Northwind.Data;
 using ITS.Vendrame.IITS.Northwind.Data.Models;
@@ -28,10 +29,15 @@ namespace ITS.Vendrame.IITS.Northwind.WebApi.Controllers
         }
 
         // GET: api/Categories/5
-        [HttpGet("{id}", Name = "Get")]
-        public Category Get(int id)
+        [HttpGet("{id}")]
+        public IActionResult Get(int id)
         {
-            return _categoriesRepository.GetById(id);
+            var category = _categoriesRepository.GetById(id);
+
+            if (category == null)
+                return NotFound();
+
+            return Ok(category);
         }
 
         // POST: api/Categories
@@ -43,17 +49,29 @@ namespace ITS.Vendrame.IITS.Northwind.WebApi.Controllers
 
         // PUT: api/Categories/5
         [HttpPut("{id}")]
-        public void Put(int id,[FromBody] Category category)
+        public IActionResult Put(int id,[FromBody] Category category)
         {
+            var dbCat = _categoriesRepository.GetById(id);
+
+            if (dbCat == null)
+                return NotFound();
+
             category.CategoryId = id;
             _categoriesRepository.Update(category);
+            return NoContent();
         }
 
         // DELETE: api/ApiWithActions/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
+            var dbCat = _categoriesRepository.GetById(id);
+
+            if (dbCat == null)
+                return NotFound();
+
             _categoriesRepository.Delete(id);
+            return NoContent();
         }
     }
 }
